[IMP] project: select project manager and task assignee based on project company
Currently, users from any company could be selected as project managers or
task assignees.

This commit adds a company-based domain to restrict user selection
according to the project's company, ensuring only relevant users can be selected.

If the project's company is listed in the user's allowed companies
(`company_ids`), the user can be selected as the project manager or assignee.

task-4598852

closes odoo/odoo#205106

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>