[IMP] http_routing, website: make HTTP error templates editable
Make the templates for HTTP errors 4XX, 400, 403, 415, and 422,
editable through the website builder. These templates can be accessed
using the following route:

    `/website/http_error/<error_code>`

This functionality allows users to customize the corresponding HTTP
error pages directly from the website editor. Users can edit the page
fully using snippets and inner blocks, and they have the option to
toggle the visibility of the default error message.

In edit mode, a demo error card will be displayed in place of the
actual error.

task-4714826