[IMP] *: improve weight and volume for packages *: stock{,_delivery,_picking_batch,_fleet}
This commit adds volume and shipping_volume in packages.
It also refactors and fixes some calculations in shipping_weight
ans shipping_volume for `stock_picking` and `stock_picking_batch` and
`report_picking_cmr`.

This commit:
1- adds computed `volume` and `shipping_volume` to packages.

2- fixes the computation of the picking shipping_volume to be the
shipping_volume of the outermost package if it's set, otherwise, it uses
the package's type dimensions if set set, otherwise, it uses the volumes
of the internals of that package and so on recursively.

3- fixes the computation of picking shipping_weight to be accurate by
making it recursive as well when you have a chain of packages inside
each other. It opts out to the shipping_weight of the package if exists,
otherwise it uses the weight of the package's type if exists in addition
to the weight of the internals of that package ans so on recursively.

4- fixes the issue of calculating shipping_weight and shipping_volume
when pickings are done in both `stock_picking_batch` and
`report_picking_cmr` using `stock_package_history` to result in the same
shipping_weight and shipping_volume calculations.

5- It also moves weight related fields on `stock_package`, `stock_move`
and `stock_picking` from `stock_delivery` module to `stock` module.

Related upgrade PR: https://github.com/odoo/upgrade/pull/8749
Task-5056474